fix(tool-activity): redact credentials in raw strings
Reuse the shared diagnostic credential policy for every string leaf in persisted raw tool payloads so URLs, headers, commands, and serialized JSON do not retain embedded secrets.
